requirements.txt
google-cloud-bigquery==3.10.0
google-auth==2.26.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-decouple==3.8

# API server
fastapi==0.111.0
uvicorn[standard]==0.30.1
